# nextjs/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Proflect API",
    "version": "1.0.0",
    "description": "REST API for the Proflect platform — resume parsing, resume builder, portfolio builder, self-test, admin management, and more."
  },
  "servers": [
    { "url": "https://proflect-neo.vercel.app", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT — obtain via POST /api/v1/auth/login and pass as `Authorization: Bearer <access_token>`"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "total": { "type": "integer" }
        }
      },
      "Resume": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "filename": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "parsed", "error"] },
          "parsed_data": { "type": "object" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "JobProfile": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "required_skills": { "type": "array", "items": { "type": "string" } },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "BuilderResume": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "template_id": { "type": "string" },
          "sections": { "type": "array", "items": { "type": "object" } },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
    "/api/v1/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 },
                  "first_name": { "type": "string" },
                  "last_name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Email already registered" }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Authenticate with email and password",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful — use `access_token` as Bearer token for subsequent requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "isAdmin": { "type": "boolean" },
                    "access_token": { "type": "string", "description": "JWT — pass as Authorization: Bearer <token>" },
                    "refresh_token": { "type": "string" },
                    "expires_in": { "type": "integer", "description": "Seconds until access_token expires" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" },
          "429": { "description": "Too many attempts / account locked" }
        }
      }
    },
...
    "/api/v1/resumes": {
      "get": {
        "tags": ["Resumes"],
        "summary": "List parsed resumes (paginated)",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Paginated resume list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resumes": { "type": "array", "items": { "$ref": "#/components/schemas/Resume" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Resumes"],
        "summary": "Create a resume entry",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "filename": { "type": "string" } } } } }
        },
        "responses": {
          "201": { "description": "Resume created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Resume" } } } }
        }
      }
    },
    "/api/v1/resumes/upload": {
      "post": {
        "tags": ["Resumes"],
        "summary": "Upload and parse a resume file",
        "requestBody": {
          "required": true,
          "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "file": { "type": "string", "format": "binary", "description": "PDF or DOCX file" } } } } }
        },
        "responses": {
          "201": { "description": "Resume uploaded and parsed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Resume" } } } },
          "400": { "description": "Invalid file type or size" }
        }
      }
    },
...
    "/api/v1/jobs": {
      "get": {
        "tags": ["Job Profiles"],
        "summary": "List job profiles",
        "responses": {
          "200": { "description": "Job list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/JobProfile" } } } } }
        }
      },
      "post": {
        "tags": ["Job Profiles"],
        "summary": "Create a job profile",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobProfile" } } }
        },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobProfile" } } } } }
      }
    },
...
    "/api/v1/builder": {
      "get": {
        "tags": ["Resume Builder"],
        "summary": "List user's builder resumes",
        "responses": { "200": { "description": "Builder resume list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/BuilderResume" } } } } } }
      },
      "post": {
        "tags": ["Resume Builder"],
        "summary": "Create new builder resume",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "title": { "type": "string" }, "template_id": { "type": "string" } } } } }
        },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BuilderResume" } } } } }
      }
    },
...
    "/api/v1/builder/{id}/writing-assist": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "post": {
        "tags": ["Resume Builder"],
        "summary": "AI writing assistance for section content",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["section_type", "content"], "properties": { "section_type": { "type": "string", "enum": ["summary", "experience", "education"] }, "content": { "type": "string" } } } } }
        },
        "responses": { "200": { "description": "Improved content returned" } }
      }
    },
...
    "/api/v1/admin/credits/requests": {
      "get": {
        "tags": ["Admin — Credits"],
        "summary": "List credit requests",
        "parameters": [{ "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "approved", "rejected"] } }],
        "responses": { "200": { "description": "Requests list" } }
      }
    },
...
  }
}